refactor: extract shared test fixtures, update documentation
Extract duplicated fixture builders (~400 lines) from averagingAnalyser.spec.ts
and dataAnalysisService.spec.ts into shared test/fixtures.ts.

Update DATA_SHAPES.md: add TaskPartial entry (Implemented), update partial
definition tasks field to Array<TaskPartial>. Create teacher-facing pedagogy
doc at docs/pedagogy/data-analysis-scoring.md. Fix DataAnalysisResponseSchema
JSDoc. Update AssignmentDefinition constructor JSDoc. Promote SPEC.md status
to Implemented v1.0.
