+++ SPEC.md
@@ -2,6 +2,7 @@
 
 ## Status
 
+- **Implemented v1.0** (all sections delivered: backend `toPartialJSON()` extension, `StudentSubmissionPartialSchema` fix, `AssignmentDefinitionPartialSchema` unification, shared `TaskPartialSchema`, `dataAnalysis.zod.ts`, `AveragingAnalyser`, `DataAnalysisService` orchestrator, teacher-facing scoring documentation.)
 - Draft v1.5 (updated 2026-06-19 — reconciled two deferred divergences into v1 scope: fixed the `StudentSubmissionPartialSchema` in `classDetailService.zod.ts` (correct nested dictionary shape replacing the flat buggy one); unified `AssignmentDefinitionPartialSchema` into a single canonical schema in `assignmentDefinitionPartials.zod.ts` (removed the duplicate lenient copy in `classDetailService.zod.ts`). All production consumers verified — zero consumers of the buggy flat submission schema; the duplicate definition schema has a known, small consumer set that is updated in-line. The analyser's input schemas are simplified to import shapes from the now-correct `classDetailService.zod.ts` and the canonical `assignmentDefinitionPartials.zod.ts` instead of defining their own.)
 - Draft v1.4 (updated 2026-06-19 — simplified per planner-reviewer feedback: removed the planned `AssignmentDefinitionPartialExtendedSchema`; the analyser's `assignmentDefinitionPartials` input reuses the existing `AssignmentDefinitionPartialsResponseSchema`; renamed `TaskWeightingSummarySchema` to `TaskPartialSchema`; extracted `TaskPartialSchema` to a small shared file in the `assignmentDefinition/` domain so the two downstream partial Zod files do not have to invert the data-load layering; dateRange validation reuses the existing strict ISO-with-timezone schema; deployment is assumed to be atomic — both the backend and frontend ship in a single release, so there is no production window requiring a temporary union type that accepts both the old `null` and the new `Array<TaskPartial>` shapes).
 - Draft v1.3 (updated 2026-06-19 — added teacher-facing algorithm documentation requirement: `docs/pedagogy/data-analysis-scoring.md` with required topics, location, integration point, and ownership).
@@ -537,7 +538,7 @@ These are the **only** backend changes required. No new endpoint is needed.
 - Update `docs/developer/DATA_SHAPES.md` with the updated `AssignmentDefinitionPartial` shape once the backend lands the task-weighting extension. This includes:
   - Adding a new "Partial Task (`TaskPartial`)" entry for `{ id: string, taskWeighting: number }`, with status `Not implemented` until the backend extension ships.
   - Updating the "Partial Definition" entry to reflect the new `tasks: Array<TaskPartial>` shape (no longer `tasks: null`).
-  - Removing or rewriting the line *"Partial definitions use `tasks: null` (not `undefined` or `{}`). This explicit marker enables fail-fast behavior when tasks are accessed without proper rehydration."* — the partial-vs-full distinction is no longer encoded by `tasks === null` on the wire (the constructor-side distinction in `AssignmentDefinition._validatePartial` is unaffected, but JSON output is `[]`).
+  - Removing or rewriting the line _"Partial definitions use `tasks: null` (not `undefined` or `{}`). This explicit marker enables fail-fast behavior when tasks are accessed without proper rehydration."_ — the partial-vs-full distinction is no longer encoded by `tasks === null` on the wire (the constructor-side distinction in `AssignmentDefinition._validatePartial` is unaffected, but JSON output is `[]`).
 - Update `docs/developer/frontend/frontend-react-query-and-prefetch.md` with the new dataset's invalidation rule once the hook exists.
 - No new canonical doc is required for the analyser itself; this `SPEC.md` is the source of truth for this delivery.
 - **Teacher-facing algorithm documentation**: a standalone help-focused document must be written as part of this delivery, explaining how the averaging algorithm computes scores so that teachers can understand and trust the results. This is **not** a developer doc — it is written for teachers and must be accessible, jargon-free, and include worked examples. The document must cover at minimum:
