+++ src/frontend/src/services/dataAnalysis/analysers/averagingAnalyser.spec.ts
@@ -1,254 +1,19 @@
 import { describe, it, expect } from 'vitest';
 import { AveragingAnalyser } from './averagingAnalyser';
 import type { AveragingAnalyserInput } from '../dataAnalysis.zod';
-
-// ---------------------------------------------------------------------------
-// Constants
-// ---------------------------------------------------------------------------
-
-/** Fixed ISO timestamp used as the default `createdAt` value in fixtures. */
-const DEFAULT_CREATED_AT = '2026-01-01T00:00:00.000Z';
+import {
+  buildInput,
+  createAssignmentPartial,
+  createClassFull,
+  createDefinitionPartial,
+  createSubmission,
+  createSubmissionItem,
+  createTaskPartial,
+  DEFAULT_CREATED_AT,
+} from '../test/fixtures';
 
 /** Tolerance (decimal places) for `toBeCloseTo` floating-point comparisons. */
 const FLOAT_TOLERANCE = 10;
-
-// ---------------------------------------------------------------------------
-// Fixture builders — produce minimal structurally-valid AveragingAnalyserInput
-// ---------------------------------------------------------------------------
-
-/**
- * Build a minimal task-partial object.
- *
- * @param {string} id - The stable task identifier.
- * @param {number} [taskWeighting=1] - The task weighting factor.
- * @returns {{ id: string, taskWeighting: number }} A minimal TaskPartial-shaped object.
- */
-function createTaskPartial(id: string, taskWeighting: number = 1) {
-  return { id, taskWeighting };
-}
-
-/**
- * Build a minimal `AssignmentDefinitionPartial`-shaped object for test use.
- *
- * @param {Object} [overrides] - Optional overrides.
- * @param {string} [overrides.definitionKey=dk_algebra] - The definition key.
- * @param {number|null} [overrides.assignmentWeighting=1] - The assignment weighting.
- * @param {Array<{id: string, taskWeighting: number}>} [overrides.tasks] - Task partials.
- * @param {string} [overrides.createdAt] - ISO creation timestamp.
- * @returns {Object} A minimal assignment-definition-partial fixture.
- */
-function createDefinitionPartial(
-  overrides: Partial<{
-    definitionKey: string;
-    assignmentWeighting: number | null;
-    tasks: Array<{ id: string; taskWeighting: number }>;
-    createdAt: string;
-  }> = {}
-) {
-  const {
-    definitionKey = 'dk_algebra',
-    assignmentWeighting = 1,
-    tasks = [createTaskPartial('t_001')],
-    createdAt = DEFAULT_CREATED_AT,
-  } = overrides;
-
-  return {
-    primaryTitle: 'Test',
-    primaryTopic: 'Algebra',
-    primaryTopicKey: 'algebra',
-    yearGroupKey: 'yg-10',
-    yearGroupLabel: 'Year 10',
-    alternateTitles: [],
-    alternateTopics: [],
-    documentType: 'assignment',
-    referenceDocumentId: null,
-    templateDocumentId: null,
-    assignmentWeighting,
-    definitionKey,
-    tasks,
-    createdAt,
-    updatedAt: null,
-  };
-}
-
-/**
- * Build a minimal submission item with optional assessments.
- *
- * @param {string} taskId - The task identifier this item belongs to.
- * @param {Object<string, {score: number|'N'}>} [assessments] - Optional criterion→score map.
- * @returns {Object} A minimal StudentSubmissionItemPartial-shaped object.
- */
-function createSubmissionItem(
-  taskId: string,
-  assessments?: Record<string, { score: number | 'N' }>
-) {
-  return {
-    id: `${taskId}_item`,
-    taskId,
-    artifact: {
-      taskId,
-      role: 'student',
-      pageId: null,
-      documentId: null,
-      content: null,
-      contentHash: null,
-      metadata: undefined,
-      uid: `${taskId}_uid`,
-      type: 'document',
-    },
-    assessments,
-    feedback: undefined,
-  };
-}
-
-/**
- * Build a minimal student submission.
- *
- * @param {string} studentId - The student identifier.
- * @param {string|null} studentName - The student display name.
- * @param {string} assignmentId - The parent assignment identifier.
- * @param {Object<string, Object>} items - Dictionary of taskId→submission item.
- * @returns {Object} A minimal StudentSubmissionPartial-shaped object.
- */
-function createSubmission(
-  studentId: string,
-  studentName: string | null,
-  assignmentId: string,
-  items: Record<string, ReturnType<typeof createSubmissionItem>>
-) {
-  return {
-    studentId,
-    studentName,
-    assignmentId,
-    documentId: null,
-    items,
-    createdAt: DEFAULT_CREATED_AT,
-    updatedAt: DEFAULT_CREATED_AT,
-  };
-}
-
-/**
- * Build a minimal assignment-partial fixture.
- *
- * @param {Object} overrides - Required partial overrides plus optional fields.
- * @param {string} overrides.assignmentId - The assignment identifier.
- * @param {string} overrides.definitionKey - The definition key.
- * @param {Array<Object>} overrides.submissions - Student submissions.
- * @param {Array<{id: string, taskWeighting: number}>} overrides.tasks - Task partials.
- * @param {number|null} [overrides.assignmentWeighting=1] - The assignment weighting.
- * @param {string} [overrides.createdAt] - ISO creation timestamp.
- * @returns {Object} A minimal AssignmentPartial-shaped object.
- */
-function createAssignmentPartial(overrides: {
-  assignmentId: string;
-  definitionKey: string;
-  submissions: ReturnType<typeof createSubmission>[];
-  tasks: Array<{ id: string; taskWeighting: number }>;
-  assignmentWeighting?: number | null;
-  createdAt?: string;
-}) {
-  const {
-    assignmentId,
-    definitionKey,
-    submissions,
-    tasks,
-    assignmentWeighting = 1,
-    createdAt = DEFAULT_CREATED_AT,
-  } = overrides;
-
-  return {
-    courseId: 'course_001',
-    assignmentId,
-    assignmentName: 'Test Assignment',
-    dueDate: null,
-    lastUpdated: null,
-    createdAt,
-    documentType: 'assessment',
-    submissions,
-    assignmentDefinition: createDefinitionPartial({
-      definitionKey,
-      assignmentWeighting,
-      tasks,
-      createdAt,
-    }),
-  };
-}
-
-/**
- * Build a minimal class-full fixture.
- *
- * @param {Object} overrides - Required class overrides with optional fields.
- * @param {string} overrides.classId - The class identifier.
- * @param {string|null} [overrides.className] - The class display name.
- * @param {Array<string>} [overrides.studentIds] - List of student identifiers.
- * @param {Array<Object>} overrides.assignments - Assignment partials.
- * @returns {Object} A minimal ClassFull-shaped object.
- */
-function createClassFull(overrides: {
-  classId: string;
-  className?: string | null;
-  studentIds?: string[];
-  assignments: ReturnType<typeof createAssignmentPartial>[];
-}) {
-  const { classId, className = 'Test Class', studentIds = [], assignments } = overrides;
-
-  return {
-    classId,
-    className,
-    cohortKey: null,
-    courseLength: 1,
-    yearGroupKey: 'yg-10',
-    classOwner: null,
-    teachers: [],
-    students: studentIds.map((id) => ({ id, name: `Student ${id}`, email: `${id}@test.com` })),
-    assignments,
-    active: null,
-  };
-}
-
-/**
- * Build a minimal `AveragingAnalyserInput` from partial overrides.
- *
- * @param {Object} overrides - Input overrides.
- * @param {Array<Object>} overrides.classOverrides - Per-class override entries.
- * @param {string} overrides.classOverrides[].classId - The class identifier.
- * @param {string|null} [overrides.classOverrides[].className] - The class display name.
- * @param {Array<string>} [overrides.classOverrides[].studentIds] - Student identifiers.
- * @param {Array<Object>} overrides.classOverrides[].assignments - Assignment partials.
- * @returns {Object} A structurally valid AveragingAnalyserInput.
- */
-function buildInput(
-  overrides: Partial<AveragingAnalyserInput> & {
-    classOverrides: Array<{
-      classId: string;
-      className?: string | null;
-      studentIds?: string[];
-      assignments: ReturnType<typeof createAssignmentPartial>[];
-    }>;
-  }
-): AveragingAnalyserInput {
-  const classIds = overrides.classOverrides.map((c) => c.classId);
-  const classes = overrides.classOverrides.map((c) => createClassFull(c));
-  const allDefinitionPartials = classes.flatMap((c) =>
-    c.assignments.map((a) => a.assignmentDefinition)
-  );
-
-  // De-duplicate by definitionKey
-  const seen = new Set<string>();
-  const uniqueDefinitionPartials = allDefinitionPartials.filter((d) => {
-    const k = d.definitionKey;
-    return seen.has(k) ? false : (seen.add(k), true);
-  });
-
-  return {
-    filter: {
-      classIds,
-    },
-    classes,
-    assignmentDefinitionPartials:
-      overrides.assignmentDefinitionPartials ?? uniqueDefinitionPartials,
-  };
-}
 
 // ---------------------------------------------------------------------------
 // Helper: deep-compare numeric results with tolerance
@@ -317,27 +82,25 @@ describe('AveragingAnalyser', () => {
     // 1) Single student, single task, single criterion (accuracy only)
     // -----------------------------------------------------------------------
     it('computes correct weighted average for single student, single task, single criterion', () => {
-      const input = buildInput({
-        classOverrides: [
-          {
-            classId: 'c_001',
-            className: 'Test Class',
-            studentIds: ['s_001'],
-            assignments: [
-              createAssignmentPartial({
-                assignmentId: 'a_001',
-                definitionKey: 'dk_algebra',
-                tasks: [createTaskPartial('t_001')],
-                submissions: [
-                  createSubmission('s_001', 'Alice', 'a_001', {
-                    t_001: createSubmissionItem('t_001', { accuracy: { score: 4 } }),
-                  }),
-                ],
-              }),
-            ],
-          },
-        ],
-      });
+      const input = buildInput([
+        {
+          classId: 'c_001',
+          className: 'Test Class',
+          studentIds: ['s_001'],
+          assignments: [
+            createAssignmentPartial({
+              assignmentId: 'a_001',
+              definitionKey: 'dk_algebra',
+              tasks: [createTaskPartial('t_001')],
+              submissions: [
+                createSubmission('s_001', 'Alice', 'a_001', {
+                  t_001: createSubmissionItem('t_001', { accuracy: { score: 4 } }),
+                }),
+              ],
+            }),
+          ],
+        },
+      ]);
 
       const analyser = new AveragingAnalyser();
       const results = analyser.analyse(input);
@@ -446,47 +209,45 @@ describe('AveragingAnalyser', () => {
     it('computes per-student and per-task breakdowns correctly for multiple students and tasks', () => {
       const studentIdsInTest = ['s_001', 's_002'];
       const tasksInMultiTest = [createTaskPartial('t_001'), createTaskPartial('t_002')];
-      const input = buildInput({
-        classOverrides: [
-          {
-            classId: 'c_001',
-            studentIds: studentIdsInTest,
-            assignments: [
-              createAssignmentPartial({
-                assignmentId: 'a_001',
-                definitionKey: 'dk_algebra',
-                tasks: tasksInMultiTest,
-                submissions: [
-                  createSubmission('s_001', 'Alice', 'a_001', {
-                    t_001: createSubmissionItem('t_001', {
-                      completeness: { score: 3 },
-                      accuracy: { score: 4 },
-                      spag: { score: 5 },
-                    }),
-                    t_002: createSubmissionItem('t_002', {
-                      completeness: { score: 4 },
-                      accuracy: { score: 3 },
-                      spag: { score: 4 },
-                    }),
+      const input = buildInput([
+        {
+          classId: 'c_001',
+          studentIds: studentIdsInTest,
+          assignments: [
+            createAssignmentPartial({
+              assignmentId: 'a_001',
+              definitionKey: 'dk_algebra',
+              tasks: tasksInMultiTest,
+              submissions: [
+                createSubmission('s_001', 'Alice', 'a_001', {
+                  t_001: createSubmissionItem('t_001', {
+                    completeness: { score: 3 },
+                    accuracy: { score: 4 },
+                    spag: { score: 5 },
                   }),
-                  createSubmission('s_002', 'Bob', 'a_001', {
-                    t_001: createSubmissionItem('t_001', {
-                      completeness: { score: 5 },
-                      accuracy: { score: 5 },
-                      spag: { score: 5 },
-                    }),
-                    t_002: createSubmissionItem('t_002', {
-                      completeness: { score: 2 },
-                      accuracy: { score: 3 },
-                      spag: { score: 4 },
-                    }),
+                  t_002: createSubmissionItem('t_002', {
+                    completeness: { score: 4 },
+                    accuracy: { score: 3 },
+                    spag: { score: 4 },
                   }),
-                ],
-              }),
-            ],
-          },
-        ],
-      });
+                }),
+                createSubmission('s_002', 'Bob', 'a_001', {
+                  t_001: createSubmissionItem('t_001', {
+                    completeness: { score: 5 },
+                    accuracy: { score: 5 },
+                    spag: { score: 5 },
+                  }),
+                  t_002: createSubmissionItem('t_002', {
+                    completeness: { score: 2 },
+                    accuracy: { score: 3 },
+                    spag: { score: 4 },
+                  }),
+                }),
+              ],
+            }),
+          ],
+        },
+      ]);
 
       const analyser = new AveragingAnalyser();
       const results = analyser.analyse(input);
@@ -667,15 +428,13 @@ describe('AveragingAnalyser', () => {
     // 4) Class with no assignments
     // -----------------------------------------------------------------------
     it('returns per-class metrics all null when class has no assignments', () => {
-      const input = buildInput({
-        classOverrides: [
-          {
-            classId: 'c_001',
-            className: 'Empty Class',
-            assignments: [],
-          },
-        ],
-      });
+      const input = buildInput([
+        {
+          classId: 'c_001',
+          className: 'Empty Class',
+          assignments: [],
+        },
+      ]);
 
       const analyser = new AveragingAnalyser();
       const results = analyser.analyse(input);
@@ -715,22 +474,20 @@ describe('AveragingAnalyser', () => {
     // 5) Student with no submissions
     // -----------------------------------------------------------------------
     it('returns per-student metrics all null when student has empty items', () => {
-      const input = buildInput({
-        classOverrides: [
-          {
-            classId: 'c_001',
-            studentIds: ['s_001'],
-            assignments: [
-              createAssignmentPartial({
-                assignmentId: 'a_001',
-                definitionKey: 'dk_algebra',
-                tasks: [createTaskPartial('t_001')],
-                submissions: [createSubmission('s_001', 'Alice', 'a_001', {})],
-              }),
-            ],
-          },
-        ],
-      });
+      const input = buildInput([
+        {
+          classId: 'c_001',
+          studentIds: ['s_001'],
+          assignments: [
+            createAssignmentPartial({
+              assignmentId: 'a_001',
+              definitionKey: 'dk_algebra',
+              tasks: [createTaskPartial('t_001')],
+              submissions: [createSubmission('s_001', 'Alice', 'a_001', {})],
+            }),
+          ],
+        },
+      ]);
 
       const analyser = new AveragingAnalyser();
       const results = analyser.analyse(input);
@@ -773,28 +530,26 @@ describe('AveragingAnalyser', () => {
     // -----------------------------------------------------------------------
     it('returns per-task metrics all null when no student submitted for a task', () => {
       const tasksInNoSubmissionTest = [createTaskPartial('t_001'), createTaskPartial('t_002')];
-      const input = buildInput({
-        classOverrides: [
-          {
-            classId: 'c_001',
-            studentIds: ['s_001'],
-            assignments: [
-              createAssignmentPartial({
-                assignmentId: 'a_001',
-                definitionKey: 'dk_algebra',
-                // Two tasks defined, but only t_001 has a submission
-                tasks: tasksInNoSubmissionTest,
-                submissions: [
-                  createSubmission('s_001', 'Alice', 'a_001', {
-                    t_001: createSubmissionItem('t_001', { accuracy: { score: 4 } }),
-                    // t_002 intentionally absent
-                  }),
-                ],
-              }),
-            ],
-          },
-        ],
-      });
+      const input = buildInput([
+        {
+          classId: 'c_001',
+          studentIds: ['s_001'],
+          assignments: [
+            createAssignmentPartial({
+              assignmentId: 'a_001',
+              definitionKey: 'dk_algebra',
+              // Two tasks defined, but only t_001 has a submission
+              tasks: tasksInNoSubmissionTest,
+              submissions: [
+                createSubmission('s_001', 'Alice', 'a_001', {
+                  t_001: createSubmissionItem('t_001', { accuracy: { score: 4 } }),
+                  // t_002 intentionally absent
+                }),
+              ],
+            }),
+          ],
+        },
+      ]);
 
       const analyser = new AveragingAnalyser();
       const results = analyser.analyse(input);
@@ -844,31 +599,29 @@ describe('AveragingAnalyser', () => {
     // 7) assignmentWeighting = 0 skips the assignment
     // -----------------------------------------------------------------------
     it('skips assignment when assignmentWeighting is 0', () => {
-      const input = buildInput({
-        classOverrides: [
-          {
-            classId: 'c_001',
-            studentIds: ['s_001'],
-            assignments: [
-              createAssignmentPartial({
-                assignmentId: 'a_001',
-                definitionKey: 'dk_algebra',
-                tasks: [createTaskPartial('t_001')],
-                assignmentWeighting: 0,
-                submissions: [
-                  createSubmission('s_001', 'Alice', 'a_001', {
-                    t_001: createSubmissionItem('t_001', {
-                      completeness: { score: 5 },
-                      accuracy: { score: 5 },
-                      spag: { score: 5 },
-                    }),
+      const input = buildInput([
+        {
+          classId: 'c_001',
+          studentIds: ['s_001'],
+          assignments: [
+            createAssignmentPartial({
+              assignmentId: 'a_001',
+              definitionKey: 'dk_algebra',
+              tasks: [createTaskPartial('t_001')],
+              assignmentWeighting: 0,
+              submissions: [
+                createSubmission('s_001', 'Alice', 'a_001', {
+                  t_001: createSubmissionItem('t_001', {
+                    completeness: { score: 5 },
+                    accuracy: { score: 5 },
+                    spag: { score: 5 },
                   }),
-                ],
-              }),
-            ],
-          },
-        ],
-      });
+                }),
+              ],
+            }),
+          ],
+        },
+      ]);
 
       const analyser = new AveragingAnalyser();
       const results = analyser.analyse(input);
@@ -905,30 +658,28 @@ describe('AveragingAnalyser', () => {
     // 8) taskWeighting = 0 skips the task
     // -----------------------------------------------------------------------
     it('skips task when taskWeighting is 0', () => {
-      const input = buildInput({
-        classOverrides: [
-          {
-            classId: 'c_001',
-            studentIds: ['s_001'],
-            assignments: [
-              createAssignmentPartial({
-                assignmentId: 'a_001',
-                definitionKey: 'dk_algebra',
-                tasks: [createTaskPartial('t_001', 0)],
-                submissions: [
-                  createSubmission('s_001', 'Alice', 'a_001', {
-                    t_001: createSubmissionItem('t_001', {
-                      completeness: { score: 5 },
-                      accuracy: { score: 5 },
-                      spag: { score: 5 },
-                    }),
+      const input = buildInput([
+        {
+          classId: 'c_001',
+          studentIds: ['s_001'],
+          assignments: [
+            createAssignmentPartial({
+              assignmentId: 'a_001',
+              definitionKey: 'dk_algebra',
+              tasks: [createTaskPartial('t_001', 0)],
+              submissions: [
+                createSubmission('s_001', 'Alice', 'a_001', {
+                  t_001: createSubmissionItem('t_001', {
+                    completeness: { score: 5 },
+                    accuracy: { score: 5 },
+                    spag: { score: 5 },
                   }),
-                ],
-              }),
-            ],
-          },
-        ],
-      });
+                }),
+              ],
+            }),
+          ],
+        },
+      ]);
 
       const analyser = new AveragingAnalyser();
       const results = analyser.analyse(input);
@@ -965,30 +716,28 @@ describe('AveragingAnalyser', () => {
     // 9) SPaG score 'N' — excluded from spag weighted sum, adjusts total weight
     // -----------------------------------------------------------------------
     it('excludes SPaG N from spag weighted sum and adjusts overall denominator', () => {
-      const input = buildInput({
-        classOverrides: [
-          {
-            classId: 'c_001',
-            studentIds: ['s_001'],
-            assignments: [
-              createAssignmentPartial({
-                assignmentId: 'a_001',
-                definitionKey: 'dk_algebra',
-                tasks: [createTaskPartial('t_001')],
-                submissions: [
-                  createSubmission('s_001', 'Alice', 'a_001', {
-                    t_001: createSubmissionItem('t_001', {
-                      completeness: { score: 3 },
-                      accuracy: { score: 4 },
-                      spag: { score: 'N' },
-                    }),
+      const input = buildInput([
+        {
+          classId: 'c_001',
+          studentIds: ['s_001'],
+          assignments: [
+            createAssignmentPartial({
+              assignmentId: 'a_001',
+              definitionKey: 'dk_algebra',
+              tasks: [createTaskPartial('t_001')],
+              submissions: [
+                createSubmission('s_001', 'Alice', 'a_001', {
+                  t_001: createSubmissionItem('t_001', {
+                    completeness: { score: 3 },
+                    accuracy: { score: 4 },
+                    spag: { score: 'N' },
                   }),
-                ],
-              }),
-            ],
-          },
-        ],
-      });
+                }),
+              ],
+            }),
+          ],
+        },
+      ]);
 
       const analyser = new AveragingAnalyser();
       const results = analyser.analyse(input);
@@ -1037,27 +786,25 @@ describe('AveragingAnalyser', () => {
       // doubleAssignmentWeighting × tripleTaskWeighting → weight = 6
       // Single data point: accuracy = 4
       // Weighted sum = 6 * 4 = 24, totalWeight = 6, value = 24 / 6 = 4
-      const input = buildInput({
-        classOverrides: [
-          {
-            classId: 'c_001',
-            studentIds: ['s_001'],
-            assignments: [
-              createAssignmentPartial({
-                assignmentId: 'a_001',
-                definitionKey: 'dk_algebra',
-                assignmentWeighting: doubleAssignmentWeighting,
-                tasks: [createTaskPartial('t_001', tripleTaskWeighting)],
-                submissions: [
-                  createSubmission('s_001', 'Alice', 'a_001', {
-                    t_001: createSubmissionItem('t_001', { accuracy: { score: 4 } }),
-                  }),
-                ],
-              }),
-            ],
-          },
-        ],
-      });
+      const input = buildInput([
+        {
+          classId: 'c_001',
+          studentIds: ['s_001'],
+          assignments: [
+            createAssignmentPartial({
+              assignmentId: 'a_001',
+              definitionKey: 'dk_algebra',
+              assignmentWeighting: doubleAssignmentWeighting,
+              tasks: [createTaskPartial('t_001', tripleTaskWeighting)],
+              submissions: [
+                createSubmission('s_001', 'Alice', 'a_001', {
+                  t_001: createSubmissionItem('t_001', { accuracy: { score: 4 } }),
+                }),
+              ],
+            }),
+          ],
+        },
+      ]);
 
       const analyser = new AveragingAnalyser();
       const results = analyser.analyse(input);
@@ -1082,30 +829,28 @@ describe('AveragingAnalyser', () => {
     // 11) Constructor defaults match explicit default weightings
     // -----------------------------------------------------------------------
     it('produces identical results between default and explicit weightings', () => {
-      const input = buildInput({
-        classOverrides: [
-          {
-            classId: 'c_001',
-            studentIds: ['s_001'],
-            assignments: [
-              createAssignmentPartial({
-                assignmentId: 'a_001',
-                definitionKey: 'dk_algebra',
-                tasks: [createTaskPartial('t_001')],
-                submissions: [
-                  createSubmission('s_001', 'Alice', 'a_001', {
-                    t_001: createSubmissionItem('t_001', {
-                      completeness: { score: 3 },
-                      accuracy: { score: 4 },
-                      spag: { score: 5 },
-                    }),
+      const input = buildInput([
+        {
+          classId: 'c_001',
+          studentIds: ['s_001'],
+          assignments: [
+            createAssignmentPartial({
+              assignmentId: 'a_001',
+              definitionKey: 'dk_algebra',
+              tasks: [createTaskPartial('t_001')],
+              submissions: [
+                createSubmission('s_001', 'Alice', 'a_001', {
+                  t_001: createSubmissionItem('t_001', {
+                    completeness: { score: 3 },
+                    accuracy: { score: 4 },
+                    spag: { score: 5 },
                   }),
-                ],
-              }),
-            ],
-          },
-        ],
-      });
+                }),
+              ],
+            }),
+          ],
+        },
+      ]);
 
       const defaultAnalyser = new AveragingAnalyser();
       const explicitAnalyser = new AveragingAnalyser({
@@ -1127,30 +872,28 @@ describe('AveragingAnalyser', () => {
       // Custom: completeness=0.6, accuracy=0.3, spag=0.1
       // Single data point: completeness=3, accuracy=4, spag=5
       // overall_i = (0.6*3 + 0.3*4 + 0.1*5) / (0.6+0.3+0.1) = (1.8+1.2+0.5)/1.0 = 3.5
-      const input = buildInput({
-        classOverrides: [
-          {
-            classId: 'c_001',
-            studentIds: ['s_001'],
-            assignments: [
-              createAssignmentPartial({
-                assignmentId: 'a_001',
-                definitionKey: 'dk_algebra',
-                tasks: [createTaskPartial('t_001')],
-                submissions: [
-                  createSubmission('s_001', 'Alice', 'a_001', {
-                    t_001: createSubmissionItem('t_001', {
-                      completeness: { score: 3 },
-                      accuracy: { score: 4 },
-                      spag: { score: 5 },
-                    }),
+      const input = buildInput([
+        {
+          classId: 'c_001',
+          studentIds: ['s_001'],
+          assignments: [
+            createAssignmentPartial({
+              assignmentId: 'a_001',
+              definitionKey: 'dk_algebra',
+              tasks: [createTaskPartial('t_001')],
+              submissions: [
+                createSubmission('s_001', 'Alice', 'a_001', {
+                  t_001: createSubmissionItem('t_001', {
+                    completeness: { score: 3 },
+                    accuracy: { score: 4 },
+                    spag: { score: 5 },
                   }),
-                ],
-              }),
-            ],
-          },
-        ],
-      });
+                }),
+              ],
+            }),
+          ],
+        },
+      ]);
 
       const analyser = new AveragingAnalyser({ completeness: 0.6, accuracy: 0.3, spag: 0.1 });
       const results = analyser.analyse(input);
@@ -1190,34 +933,32 @@ describe('AveragingAnalyser', () => {
     // 13) MetricResultSchema invariant: value=null iff applicableDataPoints=0
     // -----------------------------------------------------------------------
     it('maintains MetricResultSchema invariant throughout output', () => {
-      const input = buildInput({
-        classOverrides: [
-          {
-            classId: 'c_001',
-            studentIds: ['s_001', 's_002'],
-            assignments: [
-              createAssignmentPartial({
-                assignmentId: 'a_001',
-                definitionKey: 'dk_algebra',
-                tasks: [createTaskPartial('t_001'), createTaskPartial('t_002')],
-                submissions: [
-                  createSubmission('s_001', 'Alice', 'a_001', {
-                    t_001: createSubmissionItem('t_001', {
-                      completeness: { score: 3 },
-                      accuracy: { score: 4 },
-                      spag: { score: 5 },
-                    }),
-                    // s_001 missing t_002
+      const input = buildInput([
+        {
+          classId: 'c_001',
+          studentIds: ['s_001', 's_002'],
+          assignments: [
+            createAssignmentPartial({
+              assignmentId: 'a_001',
+              definitionKey: 'dk_algebra',
+              tasks: [createTaskPartial('t_001'), createTaskPartial('t_002')],
+              submissions: [
+                createSubmission('s_001', 'Alice', 'a_001', {
+                  t_001: createSubmissionItem('t_001', {
+                    completeness: { score: 3 },
+                    accuracy: { score: 4 },
+                    spag: { score: 5 },
                   }),
-                  createSubmission('s_002', 'Bob', 'a_001', {
-                    // s_002 has no items at all
-                  }),
-                ],
-              }),
-            ],
-          },
-        ],
-      });
+                  // s_001 missing t_002
+                }),
+                createSubmission('s_002', 'Bob', 'a_001', {
+                  // s_002 has no items at all
+                }),
+              ],
+            }),
+          ],
+        },
+      ]);
 
       const analyser = new AveragingAnalyser();
       const results = analyser.analyse(input);
@@ -1248,26 +989,24 @@ describe('AveragingAnalyser', () => {
     // 14) AppliedCriterionWeightings echoes constructor weightings
     // -----------------------------------------------------------------------
     it('echoes constructor weightings in appliedCriterionWeightings output', () => {
-      const input = buildInput({
-        classOverrides: [
-          {
-            classId: 'c_001',
-            studentIds: ['s_001'],
-            assignments: [
-              createAssignmentPartial({
-                assignmentId: 'a_001',
-                definitionKey: 'dk_algebra',
-                tasks: [createTaskPartial('t_001')],
-                submissions: [
-                  createSubmission('s_001', 'Alice', 'a_001', {
-                    t_001: createSubmissionItem('t_001', { accuracy: { score: 4 } }),
-                  }),
-                ],
-              }),
-            ],
-          },
-        ],
-      });
+      const input = buildInput([
+        {
+          classId: 'c_001',
+          studentIds: ['s_001'],
+          assignments: [
+            createAssignmentPartial({
+              assignmentId: 'a_001',
+              definitionKey: 'dk_algebra',
+              tasks: [createTaskPartial('t_001')],
+              submissions: [
+                createSubmission('s_001', 'Alice', 'a_001', {
+                  t_001: createSubmissionItem('t_001', { accuracy: { score: 4 } }),
+                }),
+              ],
+            }),
+          ],
+        },
+      ]);
 
       // Default analyser
       const defaultAnalyser = new AveragingAnalyser();
@@ -1336,9 +1075,7 @@ describe('AveragingAnalyser', () => {
           ],
         },
       ];
-      const input = buildInput({
-        classOverrides: classOverridesInSortTest,
-      });
+      const input = buildInput(classOverridesInSortTest);
 
       const analyser = new AveragingAnalyser();
       const results = analyser.analyse(input);
@@ -1355,37 +1092,35 @@ describe('AveragingAnalyser', () => {
       const zebraTasks = [createTaskPartial('t_002'), createTaskPartial('t_001')];
       const alphaTasks = [createTaskPartial('t_001')];
       const totalTaskCount = zebraTasks.length + alphaTasks.length;
-      const input = buildInput({
-        classOverrides: [
-          {
-            classId: 'c_001',
-            studentIds: ['s_001'],
-            assignments: [
-              createAssignmentPartial({
-                assignmentId: 'a_001',
-                definitionKey: 'dk_zebra',
-                tasks: zebraTasks,
-                submissions: [
-                  createSubmission('s_001', 'Alice', 'a_001', {
-                    t_001: createSubmissionItem('t_001', { accuracy: { score: 4 } }),
-                    t_002: createSubmissionItem('t_002', { accuracy: { score: 3 } }),
-                  }),
-                ],
-              }),
-              createAssignmentPartial({
-                assignmentId: 'a_002',
-                definitionKey: 'dk_alpha',
-                tasks: alphaTasks,
-                submissions: [
-                  createSubmission('s_001', 'Alice', 'a_002', {
-                    t_001: createSubmissionItem('t_001', { accuracy: { score: 5 } }),
-                  }),
-                ],
-              }),
-            ],
-          },
-        ],
-      });
+      const input = buildInput([
+        {
+          classId: 'c_001',
+          studentIds: ['s_001'],
+          assignments: [
+            createAssignmentPartial({
+              assignmentId: 'a_001',
+              definitionKey: 'dk_zebra',
+              tasks: zebraTasks,
+              submissions: [
+                createSubmission('s_001', 'Alice', 'a_001', {
+                  t_001: createSubmissionItem('t_001', { accuracy: { score: 4 } }),
+                  t_002: createSubmissionItem('t_002', { accuracy: { score: 3 } }),
+                }),
+              ],
+            }),
+            createAssignmentPartial({
+              assignmentId: 'a_002',
+              definitionKey: 'dk_alpha',
+              tasks: alphaTasks,
+              submissions: [
+                createSubmission('s_001', 'Alice', 'a_002', {
+                  t_001: createSubmissionItem('t_001', { accuracy: { score: 5 } }),
+                }),
+              ],
+            }),
+          ],
+        },
+      ]);
 
       const analyser = new AveragingAnalyser();
       const results = analyser.analyse(input);
@@ -1597,8 +1332,8 @@ describe('AveragingAnalyser', () => {
     // -----------------------------------------------------------------------
     it('resolves taskWeighting from pre-fetched assignmentDefinitionPartials cross-reference', () => {
       const preFetchedTaskWeighting = 5;
-      const input = buildInput({
-        classOverrides: [
+      const input = buildInput(
+        [
           {
             classId: 'c_001',
             studentIds: ['s_001'],
@@ -1616,14 +1351,16 @@ describe('AveragingAnalyser', () => {
             ],
           },
         ],
-        // Pre-fetched partials have taskWeighting 5 for t_001 — authoritative source
-        assignmentDefinitionPartials: [
-          createDefinitionPartial({
-            definitionKey: 'dk_algebra',
-            tasks: [createTaskPartial('t_001', preFetchedTaskWeighting)],
-          }),
-        ],
-      });
+        {
+          // Pre-fetched partials have taskWeighting 5 for t_001 — authoritative source
+          assignmentDefinitionPartials: [
+            createDefinitionPartial({
+              definitionKey: 'dk_algebra',
+              tasks: [createTaskPartial('t_001', preFetchedTaskWeighting)],
+            }),
+          ],
+        }
+      );
 
       const analyser = new AveragingAnalyser();
       const results = analyser.analyse(input);
@@ -1642,8 +1379,8 @@ describe('AveragingAnalyser', () => {
     // 21) Task weighting fallback to 1 when no matching task entry is found
     // -----------------------------------------------------------------------
     it('falls back to taskWeighting 1 when no matching task entry is found in assignmentDefinitionPartials', () => {
-      const input = buildInput({
-        classOverrides: [
+      const input = buildInput(
+        [
           {
             classId: 'c_001',
             studentIds: ['s_001'],
@@ -1661,14 +1398,16 @@ describe('AveragingAnalyser', () => {
             ],
           },
         ],
-        // Pre-fetched partials exist but have empty tasks array — no t_001 entry
-        assignmentDefinitionPartials: [
-          createDefinitionPartial({
-            definitionKey: 'dk_algebra',
-            tasks: [],
-          }),
-        ],
-      });
+        {
+          // Pre-fetched partials exist but have empty tasks array — no t_001 entry
+          assignmentDefinitionPartials: [
+            createDefinitionPartial({
+              definitionKey: 'dk_algebra',
+              tasks: [],
+            }),
+          ],
+        }
+      );
 
       const analyser = new AveragingAnalyser();
       const results = analyser.analyse(input);
@@ -1688,27 +1427,25 @@ describe('AveragingAnalyser', () => {
     // -----------------------------------------------------------------------
     it('treats null assignmentWeighting as 1', () => {
       const taskWeightingForNullTest = 2;
-      const input = buildInput({
-        classOverrides: [
-          {
-            classId: 'c_001',
-            studentIds: ['s_001'],
-            assignments: [
-              createAssignmentPartial({
-                assignmentId: 'a_001',
-                definitionKey: 'dk_algebra',
-                assignmentWeighting: null,
-                tasks: [createTaskPartial('t_001', taskWeightingForNullTest)],
-                submissions: [
-                  createSubmission('s_001', 'Alice', 'a_001', {
-                    t_001: createSubmissionItem('t_001', { accuracy: { score: 4 } }),
-                  }),
-                ],
-              }),
-            ],
-          },
-        ],
-      });
+      const input = buildInput([
+        {
+          classId: 'c_001',
+          studentIds: ['s_001'],
+          assignments: [
+            createAssignmentPartial({
+              assignmentId: 'a_001',
+              definitionKey: 'dk_algebra',
+              assignmentWeighting: null,
+              tasks: [createTaskPartial('t_001', taskWeightingForNullTest)],
+              submissions: [
+                createSubmission('s_001', 'Alice', 'a_001', {
+                  t_001: createSubmissionItem('t_001', { accuracy: { score: 4 } }),
+                }),
+              ],
+            }),
+          ],
+        },
+      ]);
 
       const analyser = new AveragingAnalyser();
       const results = analyser.analyse(input);
@@ -1835,30 +1572,28 @@ describe('AveragingAnalyser', () => {
     // 25) All criteria 'N' → overall null for that data point
     // -----------------------------------------------------------------------
     it('returns null overall when all criteria are N for a data point', () => {
-      const input = buildInput({
-        classOverrides: [
-          {
-            classId: 'c_001',
-            studentIds: ['s_001'],
-            assignments: [
-              createAssignmentPartial({
-                assignmentId: 'a_001',
-                definitionKey: 'dk_algebra',
-                tasks: [createTaskPartial('t_001')],
-                submissions: [
-                  createSubmission('s_001', 'Alice', 'a_001', {
-                    t_001: createSubmissionItem('t_001', {
-                      completeness: { score: 'N' },
-                      accuracy: { score: 'N' },
-                      spag: { score: 'N' },
-                    }),
+      const input = buildInput([
+        {
+          classId: 'c_001',
+          studentIds: ['s_001'],
+          assignments: [
+            createAssignmentPartial({
+              assignmentId: 'a_001',
+              definitionKey: 'dk_algebra',
+              tasks: [createTaskPartial('t_001')],
+              submissions: [
+                createSubmission('s_001', 'Alice', 'a_001', {
+                  t_001: createSubmissionItem('t_001', {
+                    completeness: { score: 'N' },
+                    accuracy: { score: 'N' },
+                    spag: { score: 'N' },
                   }),
-                ],
-              }),
-            ],
-          },
-        ],
-      });
+                }),
+              ],
+            }),
+          ],
+        },
+      ]);
 
       const analyser = new AveragingAnalyser();
       const results = analyser.analyse(input);
